It seems that Sudoku started to work.
I will play with it some more to see how the DPLL algorithm reduces the
number of clauses while solving the puzzle.

I've spent to much time on implementing the "or" and "not" functions in
the formula class: they are not needed since the transformation of the
sudoku to a CNF form is so detailed that it can be done directly to CNF.

I think the really hard part was to create the instrcutions on how to
convert a Sudoku in to a SAT/CNF problem not the prorgramming itslef.

The FormulaTests are not ready yet: I do not have an idea how to do
those tests.